+++ CHANGELOG.md
@@ -34,6 +34,7 @@ This release significantly revises the extraction mechanism, which means we can 
 - Updated KClause binding kextractor, which enables tristate encoding in KClause-extracted formulas
 - Extracted FeatJAR module into dedicated repository, which simplifies building and testing it standalone
 - Updated FeatJAR to latest version
+- Simplified transformation stage naming and added stage helpers for CNF transformation
 
 ### Removed